[clang-tidy] Enable bugprone-implicit-widening-of-multiplication-result
Enable the check in .clang-tidy and resolve all existing violations so it
gates new code.

The check flags int*int (or narrower) multiplications whose result is
implicitly widened to a larger type, where the product could overflow in the
narrow type before the widening. Each existing site is resolved by widening
one operand (static_cast<size_t>/<int64_t>/<uint64_t>) so the multiplication
is performed in the target type.

Seven sites where the multiplication lives inside a CHIPConfig.h macro
(CHIP_CONFIG_CASE_SESSION_RESUME_CACHE_SIZE and the
CHIP_IM_SERVER_MAX_NUM_PATH_GROUPS_* / CHIP_IM_MAX_NUM_SUBSCRIPTIONS family)
are suppressed with NOLINT at the use site: the operands are small
compile-time constants that cannot overflow, and widening the shared macro
definition would change its type tree-wide.

+++ src/app/clusters/tls-certificate-management-server/CertificateTableImpl.cpp
@@ -65,18 +65,18 @@ enum class CertificateType : uint8_t
     kRoot
 };
 
-static constexpr size_t kPersistentBufferNextIdBytes =
-    EstimateStructOverhead(sizeof(uint16_t), // mNextClientId
-                           sizeof(uint16_t), // mNextRootId,
-                           EstimateStructOverhead(sizeof(CertificateId), sizeof(FabricIndex)) *
-                               (kMaxRootCertificatesPerFabric * CHIP_CONFIG_MAX_FABRICS), // mRootCertMapping
-                           EstimateStructOverhead(sizeof(CertificateId), sizeof(FabricIndex)) *
-                               (kMaxClientCertificatesPerFabric * CHIP_CONFIG_MAX_FABRICS)); // mClientCertMapping
+static constexpr size_t kPersistentBufferNextIdBytes = EstimateStructOverhead(
+    sizeof(uint16_t), // mNextClientId
+    sizeof(uint16_t), // mNextRootId,
+    EstimateStructOverhead(sizeof(CertificateId), sizeof(FabricIndex)) *
+        (static_cast<size_t>(kMaxRootCertificatesPerFabric) * CHIP_CONFIG_MAX_FABRICS), // mRootCertMapping
+    EstimateStructOverhead(sizeof(CertificateId), sizeof(FabricIndex)) *
+        (static_cast<size_t>(kMaxClientCertificatesPerFabric) * CHIP_CONFIG_MAX_FABRICS)); // mClientCertMapping
 
 class GlobalCertificateData : public PersistableData<kPersistentBufferNextIdBytes>
 {
-    IncrementingIdHelper<CertificateId, kMaxRootCertificatesPerFabric * CHIP_CONFIG_MAX_FABRICS> mRoot;
-    IncrementingIdHelper<CertificateId, kMaxClientCertificatesPerFabric * CHIP_CONFIG_MAX_FABRICS> mClient;
+    IncrementingIdHelper<CertificateId, static_cast<size_t>(kMaxRootCertificatesPerFabric) * CHIP_CONFIG_MAX_FABRICS> mRoot;
+    IncrementingIdHelper<CertificateId, static_cast<size_t>(kMaxClientCertificatesPerFabric) * CHIP_CONFIG_MAX_FABRICS> mClient;
     EndpointId mEndpointId = kInvalidEndpointId;
 
 public:
